Configs updated

+++ .bin/picom.sh
@@ -1,10 +1,7 @@
 #!/bin/zsh
 
-POLYBAR_BG=$(xrdb -query -all | rg '\*color0' | cut -f2)
-POLYBAR_BG="#cc${POLYBAR_BG[2,7]}"
-export POLYBAR_BG
-
-killall -qw picom
-picom --daemon --experimental-backends
-
-~/.bin/polybar.sh
+if pgrep -u $UID -x picom >/dev/null; then
+    killall -qw picom
+else
+    picom --daemon --experimental-backends
+fi
